+++ src/onegov.swissvotes/onegov/swissvotes/models/__init__.py
@@ -1,4 +1,5 @@
 from onegov.swissvotes.models.actor import Actor
+from onegov.swissvotes.models.canton import Canton
 from onegov.swissvotes.models.page import TranslatablePage
 from onegov.swissvotes.models.policy_area import PolicyArea
 from onegov.swissvotes.models.principal import Principal
@@ -8,6 +9,7 @@ from onegov.swissvotes.models.vote import SwissVoteFile
 
 __all__ = (
     'Actor',
+    'Canton',
     'PolicyArea',
     'Principal',
     'SwissVote',