+++ src/onegov.swissvotes/onegov/swissvotes/models/vote.py
@@ -6,6 +6,7 @@ from onegov.file import AssociatedFiles
 from onegov.file import File
 from onegov.swissvotes import _
 from onegov.swissvotes.models.actor import Actor
+from onegov.swissvotes.models.canton import Canton
 from onegov.swissvotes.models.localized_file import LocalizedFile
 from onegov.swissvotes.models.policy_area import PolicyArea
 from sqlalchemy import Column
@@ -190,6 +191,8 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_cantons_nays = Column(Numeric(3, 1))
     result_cantons_yeas_p = Column(Numeric(13, 10))
 
+    _result_ag_accepted = Column('result_ag_accepted', Integer)
+    result_ag_accepted = encoded_property()
     result_ag_eligible_voters = deferred(Column(Integer), group='dataset')
     result_ag_votes_valid = deferred(Column(Integer), group='dataset')
     result_ag_votes_total = deferred(Column(Integer), group='dataset')
@@ -197,10 +200,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_ag_yeas = deferred(Column(Integer), group='dataset')
     result_ag_nays = deferred(Column(Integer), group='dataset')
     result_ag_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_ag_accepted = deferred(Column('result_ag_accepted', Integer),
-                                   group='dataset')
-    result_ag_accepted = encoded_property()
 
+    _result_ai_accepted = Column('result_ai_accepted', Integer)
+    result_ai_accepted = encoded_property()
     result_ai_eligible_voters = deferred(Column(Integer), group='dataset')
     result_ai_votes_valid = deferred(Column(Integer), group='dataset')
     result_ai_votes_total = deferred(Column(Integer), group='dataset')
@@ -208,10 +210,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_ai_yeas = deferred(Column(Integer), group='dataset')
     result_ai_nays = deferred(Column(Integer), group='dataset')
     result_ai_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_ai_accepted = deferred(Column('result_ai_accepted', Integer),
-                                   group='dataset')
-    result_ai_accepted = encoded_property()
 
+    _result_ar_accepted = Column('result_ar_accepted', Integer)
+    result_ar_accepted = encoded_property()
     result_ar_eligible_voters = deferred(Column(Integer), group='dataset')
     result_ar_votes_valid = deferred(Column(Integer), group='dataset')
     result_ar_votes_total = deferred(Column(Integer), group='dataset')
@@ -219,10 +220,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_ar_yeas = deferred(Column(Integer), group='dataset')
     result_ar_nays = deferred(Column(Integer), group='dataset')
     result_ar_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_ar_accepted = deferred(Column('result_ar_accepted', Integer),
-                                   group='dataset')
-    result_ar_accepted = encoded_property()
 
+    _result_be_accepted = Column('result_be_accepted', Integer)
+    result_be_accepted = encoded_property()
     result_be_eligible_voters = deferred(Column(Integer), group='dataset')
     result_be_votes_valid = deferred(Column(Integer), group='dataset')
     result_be_votes_total = deferred(Column(Integer), group='dataset')
@@ -230,10 +230,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_be_yeas = deferred(Column(Integer), group='dataset')
     result_be_nays = deferred(Column(Integer), group='dataset')
     result_be_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_be_accepted = deferred(Column('result_be_accepted', Integer),
-                                   group='dataset')
-    result_be_accepted = encoded_property()
 
+    _result_bl_accepted = Column('result_bl_accepted', Integer)
+    result_bl_accepted = encoded_property()
     result_bl_eligible_voters = deferred(Column(Integer), group='dataset')
     result_bl_votes_valid = deferred(Column(Integer), group='dataset')
     result_bl_votes_total = deferred(Column(Integer), group='dataset')
@@ -241,10 +240,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_bl_yeas = deferred(Column(Integer), group='dataset')
     result_bl_nays = deferred(Column(Integer), group='dataset')
     result_bl_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_bl_accepted = deferred(Column('result_bl_accepted', Integer),
-                                   group='dataset')
-    result_bl_accepted = encoded_property()
 
+    _result_bs_accepted = Column('result_bs_accepted', Integer)
+    result_bs_accepted = encoded_property()
     result_bs_eligible_voters = deferred(Column(Integer), group='dataset')
     result_bs_votes_valid = deferred(Column(Integer), group='dataset')
     result_bs_votes_total = deferred(Column(Integer), group='dataset')
@@ -252,10 +250,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_bs_yeas = deferred(Column(Integer), group='dataset')
     result_bs_nays = deferred(Column(Integer), group='dataset')
     result_bs_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_bs_accepted = deferred(Column('result_bs_accepted', Integer),
-                                   group='dataset')
-    result_bs_accepted = encoded_property()
 
+    _result_fr_accepted = Column('result_fr_accepted', Integer)
+    result_fr_accepted = encoded_property()
     result_fr_eligible_voters = deferred(Column(Integer), group='dataset')
     result_fr_votes_valid = deferred(Column(Integer), group='dataset')
     result_fr_votes_total = deferred(Column(Integer), group='dataset')
@@ -263,10 +260,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_fr_yeas = deferred(Column(Integer), group='dataset')
     result_fr_nays = deferred(Column(Integer), group='dataset')
     result_fr_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_fr_accepted = deferred(Column('result_fr_accepted', Integer),
-                                   group='dataset')
-    result_fr_accepted = encoded_property()
 
+    _result_ge_accepted = Column('result_ge_accepted', Integer)
+    result_ge_accepted = encoded_property()
     result_ge_eligible_voters = deferred(Column(Integer), group='dataset')
     result_ge_votes_valid = deferred(Column(Integer), group='dataset')
     result_ge_votes_total = deferred(Column(Integer), group='dataset')
@@ -274,10 +270,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_ge_yeas = deferred(Column(Integer), group='dataset')
     result_ge_nays = deferred(Column(Integer), group='dataset')
     result_ge_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_ge_accepted = deferred(Column('result_ge_accepted', Integer),
-                                   group='dataset')
-    result_ge_accepted = encoded_property()
 
+    _result_gl_accepted = Column('result_gl_accepted', Integer)
+    result_gl_accepted = encoded_property()
     result_gl_eligible_voters = deferred(Column(Integer), group='dataset')
     result_gl_votes_valid = deferred(Column(Integer), group='dataset')
     result_gl_votes_total = deferred(Column(Integer), group='dataset')
@@ -285,10 +280,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_gl_yeas = deferred(Column(Integer), group='dataset')
     result_gl_nays = deferred(Column(Integer), group='dataset')
     result_gl_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_gl_accepted = deferred(Column('result_gl_accepted', Integer),
-                                   group='dataset')
-    result_gl_accepted = encoded_property()
 
+    _result_gr_accepted = Column('result_gr_accepted', Integer)
+    result_gr_accepted = encoded_property()
     result_gr_eligible_voters = deferred(Column(Integer), group='dataset')
     result_gr_votes_valid = deferred(Column(Integer), group='dataset')
     result_gr_votes_total = deferred(Column(Integer), group='dataset')
@@ -296,10 +290,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_gr_yeas = deferred(Column(Integer), group='dataset')
     result_gr_nays = deferred(Column(Integer), group='dataset')
     result_gr_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_gr_accepted = deferred(Column('result_gr_accepted', Integer),
-                                   group='dataset')
-    result_gr_accepted = encoded_property()
 
+    _result_ju_accepted = Column('result_ju_accepted', Integer)
+    result_ju_accepted = encoded_property()
     result_ju_eligible_voters = deferred(Column(Integer), group='dataset')
     result_ju_votes_valid = deferred(Column(Integer), group='dataset')
     result_ju_votes_total = deferred(Column(Integer), group='dataset')
@@ -307,10 +300,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_ju_yeas = deferred(Column(Integer), group='dataset')
     result_ju_nays = deferred(Column(Integer), group='dataset')
     result_ju_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_ju_accepted = deferred(Column('result_ju_accepted', Integer),
-                                   group='dataset')
-    result_ju_accepted = encoded_property()
 
+    _result_lu_accepted = Column('result_lu_accepted', Integer)
+    result_lu_accepted = encoded_property()
     result_lu_eligible_voters = deferred(Column(Integer), group='dataset')
     result_lu_votes_valid = deferred(Column(Integer), group='dataset')
     result_lu_votes_total = deferred(Column(Integer), group='dataset')
@@ -318,10 +310,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_lu_yeas = deferred(Column(Integer), group='dataset')
     result_lu_nays = deferred(Column(Integer), group='dataset')
     result_lu_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_lu_accepted = deferred(Column('result_lu_accepted', Integer),
-                                   group='dataset')
-    result_lu_accepted = encoded_property()
 
+    _result_ne_accepted = Column('result_ne_accepted', Integer)
+    result_ne_accepted = encoded_property()
     result_ne_eligible_voters = deferred(Column(Integer), group='dataset')
     result_ne_votes_valid = deferred(Column(Integer), group='dataset')
     result_ne_votes_total = deferred(Column(Integer), group='dataset')
@@ -329,10 +320,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_ne_yeas = deferred(Column(Integer), group='dataset')
     result_ne_nays = deferred(Column(Integer), group='dataset')
     result_ne_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_ne_accepted = deferred(Column('result_ne_accepted', Integer),
-                                   group='dataset')
-    result_ne_accepted = encoded_property()
 
+    _result_nw_accepted = Column('result_nw_accepted', Integer)
+    result_nw_accepted = encoded_property()
     result_nw_eligible_voters = deferred(Column(Integer), group='dataset')
     result_nw_votes_valid = deferred(Column(Integer), group='dataset')
     result_nw_votes_total = deferred(Column(Integer), group='dataset')
@@ -340,10 +330,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_nw_yeas = deferred(Column(Integer), group='dataset')
     result_nw_nays = deferred(Column(Integer), group='dataset')
     result_nw_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_nw_accepted = deferred(Column('result_nw_accepted', Integer),
-                                   group='dataset')
-    result_nw_accepted = encoded_property()
 
+    _result_ow_accepted = Column('result_ow_accepted', Integer)
+    result_ow_accepted = encoded_property()
     result_ow_eligible_voters = deferred(Column(Integer), group='dataset')
     result_ow_votes_valid = deferred(Column(Integer), group='dataset')
     result_ow_votes_total = deferred(Column(Integer), group='dataset')
@@ -351,10 +340,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_ow_yeas = deferred(Column(Integer), group='dataset')
     result_ow_nays = deferred(Column(Integer), group='dataset')
     result_ow_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_ow_accepted = deferred(Column('result_ow_accepted', Integer),
-                                   group='dataset')
-    result_ow_accepted = encoded_property()
 
+    _result_sg_accepted = Column('result_sg_accepted', Integer)
+    result_sg_accepted = encoded_property()
     result_sg_eligible_voters = deferred(Column(Integer), group='dataset')
     result_sg_votes_valid = deferred(Column(Integer), group='dataset')
     result_sg_votes_total = deferred(Column(Integer), group='dataset')
@@ -362,10 +350,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_sg_yeas = deferred(Column(Integer), group='dataset')
     result_sg_nays = deferred(Column(Integer), group='dataset')
     result_sg_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_sg_accepted = deferred(Column('result_sg_accepted', Integer),
-                                   group='dataset')
-    result_sg_accepted = encoded_property()
 
+    _result_sh_accepted = Column('result_sh_accepted', Integer)
+    result_sh_accepted = encoded_property()
     result_sh_eligible_voters = deferred(Column(Integer), group='dataset')
     result_sh_votes_valid = deferred(Column(Integer), group='dataset')
     result_sh_votes_total = deferred(Column(Integer), group='dataset')
@@ -373,10 +360,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_sh_yeas = deferred(Column(Integer), group='dataset')
     result_sh_nays = deferred(Column(Integer), group='dataset')
     result_sh_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_sh_accepted = deferred(Column('result_sh_accepted', Integer),
-                                   group='dataset')
-    result_sh_accepted = encoded_property()
 
+    _result_so_accepted = Column('result_so_accepted', Integer)
+    result_so_accepted = encoded_property()
     result_so_eligible_voters = deferred(Column(Integer), group='dataset')
     result_so_votes_valid = deferred(Column(Integer), group='dataset')
     result_so_votes_total = deferred(Column(Integer), group='dataset')
@@ -384,10 +370,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_so_yeas = deferred(Column(Integer), group='dataset')
     result_so_nays = deferred(Column(Integer), group='dataset')
     result_so_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_so_accepted = deferred(Column('result_so_accepted', Integer),
-                                   group='dataset')
-    result_so_accepted = encoded_property()
 
+    _result_sz_accepted = Column('result_sz_accepted', Integer)
+    result_sz_accepted = encoded_property()
     result_sz_eligible_voters = deferred(Column(Integer), group='dataset')
     result_sz_votes_valid = deferred(Column(Integer), group='dataset')
     result_sz_votes_total = deferred(Column(Integer), group='dataset')
@@ -395,10 +380,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_sz_yeas = deferred(Column(Integer), group='dataset')
     result_sz_nays = deferred(Column(Integer), group='dataset')
     result_sz_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_sz_accepted = deferred(Column('result_sz_accepted', Integer),
-                                   group='dataset')
-    result_sz_accepted = encoded_property()
 
+    _result_tg_accepted = Column('result_tg_accepted', Integer)
+    result_tg_accepted = encoded_property()
     result_tg_eligible_voters = deferred(Column(Integer), group='dataset')
     result_tg_votes_valid = deferred(Column(Integer), group='dataset')
     result_tg_votes_total = deferred(Column(Integer), group='dataset')
@@ -406,10 +390,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_tg_yeas = deferred(Column(Integer), group='dataset')
     result_tg_nays = deferred(Column(Integer), group='dataset')
     result_tg_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_tg_accepted = deferred(Column('result_tg_accepted', Integer),
-                                   group='dataset')
-    result_tg_accepted = encoded_property()
 
+    _result_ti_accepted = Column('result_ti_accepted', Integer)
+    result_ti_accepted = encoded_property()
     result_ti_eligible_voters = deferred(Column(Integer), group='dataset')
     result_ti_votes_valid = deferred(Column(Integer), group='dataset')
     result_ti_votes_total = deferred(Column(Integer), group='dataset')
@@ -417,10 +400,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_ti_yeas = deferred(Column(Integer), group='dataset')
     result_ti_nays = deferred(Column(Integer), group='dataset')
     result_ti_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_ti_accepted = deferred(Column('result_ti_accepted', Integer),
-                                   group='dataset')
-    result_ti_accepted = encoded_property()
 
+    _result_ur_accepted = Column('result_ur_accepted', Integer)
+    result_ur_accepted = encoded_property()
     result_ur_eligible_voters = deferred(Column(Integer), group='dataset')
     result_ur_votes_valid = deferred(Column(Integer), group='dataset')
     result_ur_votes_total = deferred(Column(Integer), group='dataset')
@@ -428,10 +410,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_ur_yeas = deferred(Column(Integer), group='dataset')
     result_ur_nays = deferred(Column(Integer), group='dataset')
     result_ur_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_ur_accepted = deferred(Column('result_ur_accepted', Integer),
-                                   group='dataset')
-    result_ur_accepted = encoded_property()
 
+    _result_vd_accepted = Column('result_vd_accepted', Integer)
+    result_vd_accepted = encoded_property()
     result_vd_eligible_voters = deferred(Column(Integer), group='dataset')
     result_vd_votes_valid = deferred(Column(Integer), group='dataset')
     result_vd_votes_total = deferred(Column(Integer), group='dataset')
@@ -439,10 +420,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_vd_yeas = deferred(Column(Integer), group='dataset')
     result_vd_nays = deferred(Column(Integer), group='dataset')
     result_vd_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_vd_accepted = deferred(Column('result_vd_accepted', Integer),
-                                   group='dataset')
-    result_vd_accepted = encoded_property()
 
+    _result_vs_accepted = Column('result_vs_accepted', Integer)
+    result_vs_accepted = encoded_property()
     result_vs_eligible_voters = deferred(Column(Integer), group='dataset')
     result_vs_votes_valid = deferred(Column(Integer), group='dataset')
     result_vs_votes_total = deferred(Column(Integer), group='dataset')
@@ -450,10 +430,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_vs_yeas = deferred(Column(Integer), group='dataset')
     result_vs_nays = deferred(Column(Integer), group='dataset')
     result_vs_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_vs_accepted = deferred(Column('result_vs_accepted', Integer),
-                                   group='dataset')
-    result_vs_accepted = encoded_property()
 
+    _result_zg_accepted = Column('result_zg_accepted', Integer)
+    result_zg_accepted = encoded_property()
     result_zg_eligible_voters = deferred(Column(Integer), group='dataset')
     result_zg_votes_valid = deferred(Column(Integer), group='dataset')
     result_zg_votes_total = deferred(Column(Integer), group='dataset')
@@ -461,10 +440,9 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_zg_yeas = deferred(Column(Integer), group='dataset')
     result_zg_nays = deferred(Column(Integer), group='dataset')
     result_zg_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_zg_accepted = deferred(Column('result_zg_accepted', Integer),
-                                   group='dataset')
-    result_zg_accepted = encoded_property()
 
+    _result_zh_accepted = Column('result_zh_accepted', Integer)
+    result_zh_accepted = encoded_property()
     result_zh_eligible_voters = deferred(Column(Integer), group='dataset')
     result_zh_votes_valid = deferred(Column(Integer), group='dataset')
     result_zh_votes_total = deferred(Column(Integer), group='dataset')
@@ -472,9 +450,20 @@ class SwissVote(Base, TimestampMixin, AssociatedFiles):
     result_zh_yeas = deferred(Column(Integer), group='dataset')
     result_zh_nays = deferred(Column(Integer), group='dataset')
     result_zh_yeas_p = deferred(Column(Numeric(13, 10)), group='dataset')
-    _result_zh_accepted = deferred(Column('result_zh_accepted', Integer),
-                                   group='dataset')
-    result_zh_accepted = encoded_property()
+
+    @cached_property
+    def results_cantons(self):
+        result = {}
+        for canton in Canton.abbreviations():
+            value = getattr(self, f'_result_{canton}_accepted')
+            if value is not None:
+                result.setdefault(value, []).append(Canton(canton))
+
+        codes = self.codes('result_accepted')
+        return OrderedDict([
+            (codes[key], result[key])
+            for key in sorted(result.keys())
+        ])
 
     # Authorities
     _department_in_charge = Column('department_in_charge', Integer)